adding export files of v4 and v5
V4 did not need any screw, but will need some duct tape or adhesive part to hold it in place, as it was never designed for some multiple rotations around its axis (lets name it that way).
V5 did need two screws to hold the plate in place and with that no duct tape or adhesive part is needed.